requirements.txt
requests==2.25.0
fastapi==0.62.0
uvicorn==0.13.0
retrying==1.3.3

redis-py-cluster==2.1.0
